// IonFire/src/models/user-activity/user-activity.interface.ts
export interface UserActivity {
  $key?: string;
  uid: string;
  d_suenho_descanso: string;
  //d_suenho_rem: string;
  //d_salud: string;
  d_alimento: string;
  d_yo_cuerpo: string;
  d_yo_mente: string;
  d_otros: string;
  d_trabajo: string;
  d_humanidad: string;
  d_pareja: string;
  d_fecha: string;
  uid_fecha: string;
  week: string;
}